backend/requirements.txt
# HEXA Purchase Assistant — backend
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30
aiosqlite>=0.20
pydantic>=2.10
pydantic-settings>=2.6
python-jose[cryptography]>=3.3.0
redis>=5.2
bcrypt>=4.2
google-auth>=2.35
requests>=2.32
alembic>=1.14
greenlet>=3.1

# Tests (optional)
pytest>=8.0
httpx>=0.27
sentry-sdk[fastapi]>=2.0
